Refactor create-tf-backend.sh to support Azure Backup for Blobs
This commit refactors the create-tf-backend.sh script to add support for Azure Backup for Blobs. It introduces new variables and functions to set up Azure Backup for Blobs, create a Backup Vault, backup policy, and enable backup protection for the specified storage account. The changes also include the substitution of variables in the backup-instance-template.json file to generate the backup-instance.json file.

Refactoring the script allows for the seamless integration of Azure Backup for Blobs into the existing infrastructure. This enhancement improves the backup and recovery capabilities for the storage account, ensuring data protection and availability.